feat(temporal): widen Duration to i128 + correctly-rounded total/BigInt→Number (+14)
Last concentrated Temporal code vein — Duration i128 precision + rounding.

- Widened DurationFields from i64 to i128 (the value type in temporal_iso.rs + all
  arithmetic across the six per-type modules, ~150 lossless type-site fixes). A valid
  Duration's microseconds reach ~9e21 and nanoseconds ~9e24 (2^53 seconds' worth),
  which saturated/wrapped at i64 — breaking Duration.from({nanoseconds:9e24}), add/
  subtract overflow guards, and round() balancing.
- ratio_to_f64: new round-half-even i128/i128→f64 helper; Duration.prototype.total now
  rounds the exact rational ONCE instead of the old whole+rem double-round near 2^53.
- BigInt::to_f64 (bignum.rs): route i128-fitting values through Rust's IEEE-correct
  `i128 as f64` (the puremp backend mis-rounds a tie's neighbour, e.g.
  Number(8692288669465520373761n)); backend stays the >2^127 fallback.
- 7 unit tests (ratio_to_f64 large-quotient single-rounding edges, i128 large-value
  balance_time_duration).

built-ins/Temporal/Duration 19→7; broad built-ins/Temporal 146→132 (Duration underpins
every type's until/since/round — the win rippled); intl402/Temporal 341 unchanged (0
regr); built-ins/BigInt + language/types/bigint + built-ins/Number all 0 regressions
(bignum.rs shared-file change verified); lib 890; both build configs compile.

Remaining 7 Duration fails are the relativeTo range-limit + DST/ZonedDateTime-anchored
rounding vein (different from i128 arithmetic), deferred.

+++ src/nbexec/temporal_instant.rs
@@ -447,7 +447,7 @@ impl<'a> Interp<'a> {
             if !n.is_finite() || n.fract() != 0.0 {
                 return Err(self.range_err("duration fields must be integers"));
             }
-            let iv = n as i64;
+            let iv = n as i128;
             match f {
                 "days" => d.days = iv,
                 "hours" => d.hours = iv,
